Remove mis-interpreted sticker exports (were seen as stories)
